+++ common/src/main/java/io/homo/superresolution/core/graphics/opengl/texture/GlTexture2D.java
@@ -196,7 +196,7 @@ public class GlTexture2D implements ITexture, IDebuggableObject {
 
     @Override
     public void updateDebugLabel(String newLabel) {
-        setGlObjectLabel(GL_TEXTURE, handle(), getDebugLabel());
+        setGlObjectLabel(GL_TEXTURE, (int) handle(), getDebugLabel());
     }
 
     public void generateMipmap() {
